Extract test IApiRequestFactory helpers for resuse
